Implement candidate test flow: API DTOs, repository, use cases, viewmodels, UI and navigation

+++ app/src/main/java/com/example/diplomproject/domain/model/TestSessionModels.kt
@@ -25,11 +25,11 @@ data class NextQuestionPayload(
 )
 
 data class ScaleScores(
-    val attention: Int,
-    val stressResistance: Int,
-    val responsibility: Int,
-    val adaptability: Int,
-    val decisionSpeedAccuracy: Int,
+    val attention: Double,
+    val stressResistance: Double,
+    val responsibility: Double,
+    val adaptability: Double,
+    val decisionSpeedAccuracy: Double,
 )
 
 data class CandidateResultHistoryItem(
